refactor(iroh)!: Extract net and node rpc (#2927)
## Description

Introduce a new crate iroh-node-util.

This contains generic utilities that are needed to build an iroh node,
as well as the rpc protocol and rpc client to remote control an
arbitrary iroh-net endpoint (add/remove addresses etc.) as well as to
control the node as a whole (shutdown/stats/status).

## Breaking Changes

- module iroh::client::net is now a reexport of
iroh_node_util::rpc::client::net
- module iroh::client::node is now a reexport of
iroh_node_util::rpc::client::node
- fn iroh::client::Iroh::net returns a net::Client instead of a
&net::Client
- fn iroh::client::Iroh::node returns a node::Client instead of a
&node::Client

## Notes & open questions

Note: I think the rpc protocol and the client needs some refactoring. It
isn't always clear to me what is node and what is net. But I think it
might be best to do this in a subsequent PR and keep this one as a pure
move things around PR for better separation.

## Change checklist

- [ ] Self-review.
- [ ] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [ ] Tests if relevant.
- [ ] All breaking changes documented.

---------

Co-authored-by: dignifiedquire <me@dignifiedquire.com>
Co-authored-by: Diva M <divma@protonmail.com>